pyproject.toml
[project]
name = "practice-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents[litellm]>=0.0.16",
    "pydantic>=2.11.5",
]
